requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.3
django-cors-headers==3.13.0
django-filter==22.1
django-jazzmin==2.6.0
django-rest-auth==0.9.5
django-storages==1.13.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.2.1
drf-yasg==1.21.4
email-validator==1.3.0
gunicorn==20.1.0
holidays-co==1.0.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
packaging==21.3
Pillow==9.3.0
psycopg2==2.9.5
PyJWT==2.6.0
pyparsing==3.0.9
pytz==2022.6
requests==2.28.1
ruamel.yaml==0.17.21
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.13
whitenoise==6.2.0
